requirements.txt
Flask==1.1.1
flask_dance==3.0.0
python-dotenv==0.12.0
#flask_pymongo==2.3.0
pymongo==3.10.1
dnspython==1.16.0 # needed for pymongo to use mongodb+srv:// URIs